+++ apps/explorer/config/config.exs
@@ -36,6 +36,8 @@ config :explorer, Explorer.Counters.AverageBlockTime,
 
 config :explorer, Explorer.Celo.AbiHandler, enabled: true
 
+config :explorer, Explorer.Celo.SignerCache, enabled: true
+
 config :explorer, :stacktrace_depth, 20
 
 config :explorer, Explorer.Chain.Events.Listener,
