+++ apps/explorer/lib/explorer/celo/account_reader.ex
@@ -4,11 +4,11 @@ defmodule Explorer.Celo.AccountReader do
   """
 
   require Logger
-  alias Explorer.Celo.AbiHandler
-  alias Explorer.SmartContract.Reader
+  alias Explorer.Celo.SignerCache
 
   use Bitwise
-  use Explorer.Celo.Util
+
+  import Explorer.Celo.Util
 
   def account_data(%{address: account_address}) do
     data = fetch_account_data(account_address)
@@ -266,12 +266,6 @@ defmodule Explorer.Celo.AccountReader do
     data
   end
 
-  def epoch_signers(epoch, epoch_size, block_number) do
-    case SignerCache.fetch_signers(epoch) do
-      {:ok, signers} -> signers
-      _ -> 
-  end
-
   def validator_history(block_number) do
     data = fetch_validators(block_number)
 
@@ -283,8 +277,9 @@ defmodule Explorer.Celo.AccountReader do
          {:ok, [epoch]} <- data["getEpochNumberOfBlock"],
          {:ok, gold_address} <- get_address("GoldToken"),
          {:ok, usd_address} <- get_address("StableToken"),
-         {:ok, oracle_address} <- get_address("SortedOracles"),
-         {:ok, [validators]} <- epoch_signers(epoch, epoch_size, block_number) do
+         {:ok, oracle_address} <- get_address("SortedOracles") do
+      
+      validators = SignerCache.epoch_signers(epoch, epoch_size, block_number)
       list =
         validators
         |> Enum.with_index()
@@ -323,76 +318,4 @@ defmodule Explorer.Celo.AccountReader do
     call_methods([{:lockedgold, "getPendingWithdrawals", [address]}])
   end
 
-  def call_methods(methods) do
-    contract_abi = AbiHandler.get_abi()
-
-    methods
-    |> Enum.map(&format_request/1)
-    |> Enum.filter(fn req -> req.contract_address != :error end)
-    |> Enum.map(fn %{contract_address: {:ok, address}} = req -> Map.put(req, :contract_address, address) end)
-    |> Reader.query_contracts(contract_abi)
-    |> Enum.zip(methods)
-    |> Enum.into(%{}, fn
-      {response, {_, function_name, _}} -> {function_name, response}
-      {response, {_, function_name, _, _}} -> {function_name, response}
-    end)
-  end
-
-  defp format_request({contract_name, function_name, params}) do
-    %{
-      contract_address: contract(contract_name),
-      function_name: function_name,
-      args: params
-    }
-  end
-
-  defp format_request({contract_name, function_name, params, bn}) do
-    %{
-      contract_address: contract(contract_name),
-      function_name: function_name,
-      args: params,
-      block_number: bn
-    }
-  end
-
-  defp contract(:blockchainparameters), do: get_address("BlockchainParameters")
-  defp contract(:lockedgold), do: get_address("LockedGold")
-  defp contract(:validators), do: get_address("Validators")
-  defp contract(:election), do: get_address("Election")
-  defp contract(:epochrewards), do: get_address("EpochRewards")
-  defp contract(:accounts), do: get_address("Accounts")
-  defp contract(:gold), do: get_address("GoldToken")
-  defp contract(:usd), do: get_address("StableToken")
-
-  def get_address(name) do
-    case get_address_raw(name) do
-      {:ok, address} -> {:ok, "0x" <> Base.encode16(address, case: :lower)}
-      _ -> :error
-    end
-  end
-
-  def get_address_raw(name) do
-    contract_abi = AbiHandler.get_abi()
-
-    methods = [
-      %{
-        contract_address: "0x000000000000000000000000000000000000ce10",
-        function_name: "getAddressForString",
-        args: [name]
-      }
-    ]
-
-    res =
-      methods
-      |> Reader.query_contracts(contract_abi)
-      |> Enum.zip(methods)
-      |> Enum.into(%{}, fn {response, %{function_name: function_name}} ->
-        {function_name, response}
-      end)
-
-    case res["getAddressForString"] do
-      {:ok, [address]} -> {:ok, address}
-      _ -> :error
-    end
-  end
 end
