+++ apps/indexer/lib/indexer/fetcher/internal_transaction.ex
@@ -12,7 +12,7 @@ defmodule Indexer.Fetcher.InternalTransaction do
 
   import Indexer.Block.Fetcher, only: [async_import_coin_balances: 2]
 
-  alias Explorer.Celo.AccountReader
+  alias Explorer.Celo.Util
   alias Explorer.Chain
   alias Explorer.Chain.Block
   alias Explorer.Chain.Cache.{Accounts, Blocks}
@@ -171,7 +171,7 @@ defmodule Indexer.Fetcher.InternalTransaction do
 
     # Gold token special updates
     with true <- Application.get_env(:indexer, Indexer.Block.Fetcher, [])[:enable_gold_token],
-         {:ok, gold_token} <- AccountReader.get_address("GoldToken") do
+         {:ok, gold_token} <- Util.get_address("GoldToken") do
       set = add_gold_token_balances(gold_token, addresses_params, MapSet.new())
       TokenBalance.async_fetch(MapSet.to_list(set))
     end
